refactor(openomni): merge executor satellites into executor.ts
executor-abort/dispatch/settlement each had exactly one importer
(executor.ts). executor-events.ts stays separate as the sole emitter
surface.
