--- packages/openomni/src/execution-runtime/tool/executor-dispatch.ts
@@ -1,48 +1,0 @@
-import type { Tool } from "@openomni/protocol";
-import type { ImplicitInputSource, NativeTool, ToolRuntimeContext } from "./types.js";
-
-export function buildDispatchTable(tools: readonly NativeTool[]): Map<string, NativeTool> {
-  const dispatch = new Map<string, NativeTool>();
-  for (const tool of tools) {
-    dispatch.set(tool.spec.name, tool);
-    const sanitized = tool.spec.name.replace(/\./g, "_");
-    if (sanitized !== tool.spec.name) dispatch.set(sanitized, tool);
-  }
-  return dispatch;
-}
-
-function resolveImplicitValue(
-  source: ImplicitInputSource,
-  runtime: ToolRuntimeContext,
-): string | undefined {
-  switch (source) {
-    case "sessionId":
-      return runtime.sessionId;
-    case "runId":
-      return runtime.runId;
-    case "agentName":
-      return runtime.agentName;
-    case "workspaceRoot":
-      return runtime.workspaceRoot;
-  }
-}
-
-export function injectImplicitInputs(
-  call: Tool.Call,
-  tool: NativeTool,
-  runtime: ToolRuntimeContext | undefined,
-): Tool.Call {
-  if (!tool.implicitInputs || !runtime) return call;
-
-  const injected: Record<string, unknown> = { ...call.input };
-  for (const [param, source] of Object.entries(tool.implicitInputs)) {
-    const value = resolveImplicitValue(source, runtime);
-    if (value !== undefined) injected[param] = value;
-  }
-  return { ...call, input: injected };
-}
-
-export function resolveDispatchedCall(call: Tool.Call, tool: NativeTool): Tool.Call {
-  const originalName = tool.spec.name;
-  return originalName === call.tool ? call : { ...call, tool: originalName };
-}
